+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - MediaInput backend has been reworked.
   - Now supports series.
   - Now utilizes async to parse MediaInfo of multiple files concurrently, this greatly speeds up read speeds for anything over a single file.
+- Error dialog can now be maximized and grows to it's parent geometry on error.
 
 ### Fixed